Changes in organisation following review comments.
