+++ requirements-pip.txt
@@ -1,2 +1,1 @@
 hdx-python-api==4.3.5
-snakemake==5.13.0
